# static/swaggerAuth.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API Documentation",
    "description": "API for marketplace.",
    "version": "1.0.3"
  },
  "paths": {
    "/accounts/signup": {
      "post": {
        "summary": "User Registration",
        "tags": [
          "Accounts"
        ],
        "requestBody": {
          "description": "User registration data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupUserSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful registration",
            "content": {
              "application/json": {
                "example": {
                  "user": {
                    "id": 1,
                    "email": "user@example.com",
                    "full_name": "John Doe"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "User with such email already exists"
          }
        }
      }
    },
    "/accounts/signin": {
      "post": {
        "summary": "User Sign-In",
        "tags": [
          "Accounts"
        ],
        "requestBody": {
          "description": "User sign-in data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninUserSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful sign-in",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "your_access_token",
                  "refresh_token": "your_refresh_token"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/verify/{token}": {
      "get": {
        "summary": "Verify Email",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "example": {
                  "message": "OK"
                }
              }
            }
          },
          "404": {
            "description": "Invalid verification token"
          }
        }
      }
    },
...
    "/accounts/refresh": {
      "post": {
        "summary": "Refresh Access Token",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New access token generated",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "new_access_token",
                  "refresh_token": "new_refresh_token"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/logout": {
      "delete": {
        "summary": "User Logout",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token successfully revoked",
            "content": {
              "application/json": {
                "example": {
                  "msg": "Access token successfully revoked"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/change_phone_number": {
      "post": {
        "summary": "Change Phone Number",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Phone number change data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneChangeSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phone number updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Phone number updated successfully"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/accounts/change_full_name": {
      "post": {
        "summary": "Change Full Name",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Full name change data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullNameChangeSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Full name updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Full name updated successfully"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/accounts/profile_photo": {
      "post": {
        "summary": "Upload Profile Photo",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Profile photo upload data",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePhotoUploadSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile photo uploaded successfully",
            "content": {
              "application/json": {
                "example": {
                  "profile_picture": "http://127.0.0.1:5000/static/media/profile/file_name.jpg"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "413": {
            "$ref": "#/components/responses/EntityTooLarge3"
          }
        }
      },
      "delete": {
        "summary": "Delete Profile Photo",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile photo deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Profile photo deleted successfully"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFound"
          }
        }
      }
    },
...
    "/accounts/change_password": {
      "post": {
        "summary": "Change Password",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Password change data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Password changed successfully"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/accounts/info": {
      "get": {
        "summary": "Get User Information",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "EntityTooLarge3": {
        "description": "Request entity too large",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "File size too large. Maximum file size is 3MB."
                }
              }
            }
          }
        }
      },
...
      "UnauthorizedError": {
        "description": "Unauthorized error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Missing Authorization Header"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request data"
                }
              }
            }
          }
        }
      },
...
      "UserNotFound": {
        "description": "User not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User not found"
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "SignupUserSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "full_name",
          "password"
        ]
      },
...
      "SigninUserSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "PhoneChangeSchema": {
        "type": "object",
        "properties": {
          "phone_number": {
            "type": "string"
          }
        },
        "required": [
          "phone_number"
        ]
      },
      "FullNameChangeSchema": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          }
        },
        "required": [
          "full_name"
        ]
      },
      "ProfilePhotoUploadSchema": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "image"
        ]
      },
...
      "PasswordChangeSchema": {
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "current_password",
          "new_password"
        ]
      },
...
      "UserInfoResponse": {
        "type": "object",
        "properties": {
          "phone_number": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "profile_photo": {
            "type": "string",
            "format": "uri",
            "description": "URL to the user's profile photo"
          },
          "post": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "branch_name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "shop_id": {
            "type": "integer"
          },
          "have_a_shop": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Bearer token"
      }
    }
  }
}